- Refactor and extract into hooks and components, so that we can have multiple variants.
